tracker: median flow, optical flow
objtracker:
Added Optical Flow support
The new model is a combination of optical flow and median flow tracking algorithms.
The module devise a methodology where we use both:
1) optical flow based tracking of the mid-point of a bounding box detected
in assess_iou_trackerBBs_detectedBBs(), the IoU of optical flow detected BB is assigned 1.0 (the best IoU)
when the point of interest is in the detected BB or around the BB by a tolerance factor
2) median flow based tracking of the entire bounding box

assess_iou_trackerBBs_detectedBBs() function shall take in the tracking results from both the models,
compare them with detected BBs and assign the BB-IDs
: The function calculates the IoU between actual objects in the next frame and each of the tracked bounding boxes.
The best IoU score will win and module suggests that bounding box as the tracked bounding box which is used to assign a unique BB-ID (as seen in the below shared videos).